refactor: rehash project modules
