+++ crates/canopus/src/infra.rs
@@ -1,6 +1,6 @@
 // Copyright 2025 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-mod arguments;
-
-pub use arguments::parse_arguments;
+pub mod cli;
+pub mod github;
+pub mod paths;
