+++ crates/canopus/src/main.rs
@@ -1,9 +1,11 @@
 // Copyright 2025 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-mod cli;
+use crate::infra::cli;
+
 mod core;
 mod features;
+mod infra;
 
 fn main() -> anyhow::Result<()> {
     better_panic::install();
