+++ crates/canopus/tests/component.rs
